split file for contraction and einsum
