+++ yastn/tensor/_control_lru.py
@@ -15,7 +15,7 @@
 """ Dynamical changing of lru_cache maxsize. """
 from functools import lru_cache
 
-from . import _merging, _contractions, _algebra
+from . import _algebra, _merging, _contractions, _einsum
 
 __all__ = ['clear_cache', 'get_cache_info', 'set_cache_maxsize']
 
@@ -32,7 +32,7 @@ def set_cache_maxsize(maxsize=0):
     _contractions._meta_swap_gate_charge = lru_cache(maxsize)(_contractions._meta_swap_gate_charge.__wrapped__)
     _contractions._meta_trace = lru_cache(maxsize)(_contractions._meta_trace.__wrapped__)
     _contractions._meta_vdot = lru_cache(maxsize)(_contractions._meta_vdot.__wrapped__)
-    _contractions._meta_ncon = lru_cache(maxsize)(_contractions._meta_ncon.__wrapped__)
+    _einsum._meta_ncon = lru_cache(maxsize)(_einsum._meta_ncon.__wrapped__)
     _merging._meta_merge_to_matrix = lru_cache(maxsize)(_merging._meta_merge_to_matrix.__wrapped__)
     _merging._meta_unmerge_matrix = lru_cache(maxsize)(_merging._meta_unmerge_matrix.__wrapped__)
     _merging._masks_hfs_intersection = lru_cache(maxsize)(_merging._masks_hfs_intersection.__wrapped__)
@@ -54,7 +54,7 @@ def clear_cache():
     _contractions._meta_swap_gate_charge.cache_clear()
     _contractions._meta_trace.cache_clear()
     _contractions._meta_vdot.cache_clear()
-    _contractions._meta_ncon.cache_clear()
+    _einsum._meta_ncon.cache_clear()
     _merging._meta_merge_to_matrix.cache_clear()
     _merging._meta_unmerge_matrix.cache_clear()
     _merging._masks_hfs_intersection.cache_clear()
@@ -82,5 +82,5 @@ def get_cache_info():
             "vdot": _contractions._meta_vdot.cache_info(),
             "swap_gate": _contractions._meta_swap_gate.cache_info(),
             "swap_gate_charge": _contractions._meta_swap_gate_charge.cache_info(),
-            "ncon": _contractions._meta_ncon.cache_info(),
+            "ncon": _einsum._meta_ncon.cache_info(),
             "addition": _algebra._meta_addition.cache_info()}
